Move some interfaces to Core
* added IFilterableIssue interface
