+++ src/Integration.UnitTests/Suppression/IssuesFilterTests.cs
@@ -22,10 +22,10 @@ using System;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Integration.Suppression;
-using SonarLint.VisualStudio.Integration.Vsix.SonarLintTagger;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests.SonarLintTagger
+namespace SonarLint.VisualStudio.Integration.UnitTests.Suppression
 {
     [TestClass]
     public class IssuesFilterTests
